requirements.txt
fastapi
uvicorn
pymongo
python-jose
passlib[bcrypt]
pydantic
python-dotenv
